--- internal/db/db.go
@@ -1,13 +1,0 @@
-package db
-
-import (
-	"context"
-
-	"github.com/jackc/pgx/v5/pgxpool"
-)
-
-type DB = pgxpool.Pool
-
-func New(ctx context.Context, dsn string) (*pgxpool.Pool, error) {
-	return pgxpool.New(ctx, dsn)
-}
