--- internal/domain/session.go
@@ -1,9 +1,0 @@
-package domain
-
-import "time"
-
-type Session struct {
-	SessionToken string    `json:"session_token"`
-	UserID       string    `json:"user_id"`
-	Expires      time.Time `json:"expires"`
-}
